Title: 쪼개어 배낭 채우기 구현, Time: 166ms, Memory: 1MB, Status: Runtime Error - CodeTree

+++ 240130/쪼개어 배낭 채우기 구현/implement-fractional-knapsack.cpp
@@ -15,7 +15,7 @@ int main() {
     // for(int i = 0; i < n; i++){
     //     cout << v[i].first << ' ' << v[i].second << '\n';
     // }
-    double sum = 0;
+    long double sum = 0;
     int index = n - 1;
     cout << fixed;
     cout.precision(3);
